requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-environ
django-cors-headers
celery
redis
django-celery-results
django-celery-beat
requests
eventlet
